Create TestDifferentAlgorithms
I was creating a simpler problem with the same concept to RotateNinetyDegrees from Problem Assignment 1 to test two algorithms.  These algorithms are part of getPixel().  The source algorithm is getWidth2(). I was suspicious that the source algorithm was keeping my solution from working.  This is the code I used to debug my solution so I knew how to fix it.  Whether the source algorithm is incompatible with how I attempted to solve the RotateNinetyDegrees problem or if the algorithm is not properly working at this point is unknown.  Currently the source algorithm function call is commented out. To test between the two algorithms, comment out getWidth() and run getWidth2() instead.